+++ dev/dw1000/tdoa/tdoa_tag/tdoa_tag.h
@@ -3,6 +3,6 @@
 
 #include <stdint.h>
 
-void handleTagRxPacket(uint32_t rxTime, const uint8_t *packetbuf, const uint16_t data_len);
+void handleTagRxPacket(uint64_t rxTime, const uint8_t *packetbuf, const uint16_t data_len);
 void tdoaTagInit();
 #endif
